# openapi.yaml
openapi: 3.0.3
info:
  title: IM Backend API
  version: 0.1.0
paths:
...
  /conversations/{conversationID}/ack:
    post:
      operationId: acknowledgeConversation
      summary: Advance one device's durable cursor for a conversation
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ConversationID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcknowledgeConversationRequest'
      responses:
        '200':
          description: Current durable device cursor
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceConversationSyncState'
        '400':
          description: Invalid conversation identifier, JSON, or cursor
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationAckBadRequest'
        '401':
          description: A valid access token is required
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationRequired'
        '404':
          description: Conversation is not visible to the authenticated user
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationNotFound'
        '409':
          description: Cursor is ahead of the current conversation stream
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursorAhead'
        '500':
          description: Unexpected acknowledgement failure
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageInternalError'
        '503':
          description: Authentication dependency is temporarily unavailable
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-token
  headers:
    RequestID:
      description: Server-generated identifier used to correlate the response with logs
      required: true
      schema:
        type: string
        minLength: 1
...
  parameters:
    ConversationID:
      name: conversationID
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  schemas:
...
    AuthenticationRequired:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          enum:
            - AUTHENTICATION_REQUIRED
        message:
          type: string
          minLength: 1
        requestId:
          type: string
          minLength: 1
...
    MessageInternalError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          enum:
            - INTERNAL_ERROR
        message:
          type: string
          minLength: 1
        requestId:
          type: string
          minLength: 1
...
    AcknowledgeConversationRequest:
      type: object
      additionalProperties: false
      required:
        - cursor
      properties:
        cursor:
          type: integer
          format: int64
          minimum: 0
    DeviceConversationSyncState:
      type: object
      additionalProperties: false
      required:
        - conversationId
        - appliedCursor
        - updatedAt
      properties:
        conversationId:
          type: integer
          format: int64
          minimum: 1
        appliedCursor:
          type: integer
          format: int64
          minimum: 0
        updatedAt:
          type: string
          format: date-time
...
    ConversationAckBadRequest:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          enum:
            - INVALID_CONVERSATION_ID
            - INVALID_JSON
            - INVALID_ACK
        message:
          type: string
          minLength: 1
        requestId:
          type: string
          minLength: 1
    ConversationNotFound:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          enum:
            - CONVERSATION_NOT_FOUND
        message:
          type: string
          minLength: 1
        requestId:
          type: string
          minLength: 1
    CursorAhead:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          enum:
            - CURSOR_AHEAD
        message:
          type: string
          minLength: 1
        requestId:
          type: string
          minLength: 1
    AuthenticationUnavailable:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          enum:
            - AUTHENTICATION_UNAVAILABLE
        message:
          type: string
          minLength: 1
        requestId:
          type: string
          minLength: 1
...